Move functions in AddFood to new component AddNewFood
